Move filesystem migration to its own module
Keep filesystem-specific code together so generic storage utilities can
compile without that backend. This commit only relocates the existing
migration behavior and tests.
